Fix xdm/rdm_qcm_setup to set all fields
The kernel communciation was getting EBUSY/ETIME errors
because of issues with the valid toggling for the XDM/RDM
queues and the queues wrapped around. Presumably because the
assumption was that the use level would initialize all the
queue indicies, include the toggle bits, the driver setup
routines left them unset. Routines were modified to initialize
all the fields.

Addditonally, there was a simulator bug in the RDM code that
meant it would always write a one for the valid bit. The following
PR fixes the problem:

https://github.hpe.com/TMAS/archsim/pull/728

Revisit: perhaps the XDM completions should be checked for
errors.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>