Return EEXIST from zhpe_umem_get()
EEXIST is being returned if we try to register the same exact
parameters that already exist; at the moment zhpe_umem_get()
finds the existing entry and reuses it, which can cause sw_pg
tree to become corrupted. So, just return EEXIST early and
we'll figure out how to handle things better later.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>