Take another crack at the clean-up problem
The previous "placing a hold" solution prevented the
simulator warnings by actually preventing the process's mm
from being cleaned up, so the driver's release function was
never called.

Use a mmu_notifier to get a call when the mm_struct is
being destroyed before the page table is freed.

At some point, we need to use the notifier to clean-up
individual mappings and notify remotes if they are
unmapped.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>