Add support for big responder zmmu entries
Code works in co-operation with library.

Moved ZHPE_MR_INDIVIDUAL flag to first byte so it would be
exported to remote side. Added a ZHPE_MR_ZMMU_ONLY flag so
the library can create the big zmmu entry. The big zmmu
entry in not a part of the umem tree, it is saved in
big_rsp_umem in the file_data; there can only be one. For
registrations without the ZHPE_MR_INDVIDUAL, the
big_rsp_umem is used to compute the rsp_zaddr. Individual
mapping should be able to co-exist, but that hasn't been
tested.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>