Fix CentOS crash due to AVX alignment issues
When building the driver against the CentOS kernel, the options
use set the stack alignment to 16 bytes, but the CentOS kernel
does not guarantee this alignment when calling the driver in all
cases, so we'll use the VMOVDQU instruction for AVX CPU memory
accesses since it isn't alignment sensitive.

We haven't seen this issue on the various 4.x kernels we've
tested, but I don't understand if that is just luck of some
kind.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>