An unfortunately large check-in to assist debugging
In order to debug, I needed to know the pid of every debug
message from the driver. So, I folded the debugx macros
into the debug debug macros and proceeded to ripple changes
to every output statement throughout the driver.

There were a couple of files where the mix of 8 and 4 space
indentation annoyed me and I just tried to slam them to
4 with emacs, since that is more forgiving. If we were
going to submit this driver to the kernel, I would care
more about forcing it all to 8.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>