Do snapshot logic only on teardown
There appears to be a low-frequency lock-up (1 in 10^6) in the
snapshot hardware. We really *should* really only need it in
the teardown case; in the other cases, the SW handshaking should
prevent the case the snapshot logic is trying to prevent. So,
far it seems to get rid of the lock-up and is not causing any
grief.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>