Rewrite driver XDM/RDM paths
Send/receive paths now approximate that of the user-level code.

On the send side, the driver now keeps a copy of every issued
command and uses a bitmap to allocate completion indicies and
the buffer slot. The tail management is radically simplified
and this allows us to find uncompleted commands when queues
hang. Also, it will ease the introduction of command buffers
and retries at some point in the future.

On the receive side, the driver attempts to reduce the number
of head updates in the case of multiple messages being received
at once.

The default values for two parameters was changed:
kmsg_timeout=10
msg_qsize=8192

Signed-off-by: John Byrne <john.l.byrne@hpe.com>